BUG4204 バグ報告① 「承認待ち」タブの一覧表示に不具合
ST#4279 承認時に支払依頼確認モーダルが開かない不具合
ST#4282 差し戻しされた支払依頼を再申請すると、承認ステータスが「差し戻し」のままになる時がある
ST#4283 最終承認後は閲覧のみにさせる

Related work items: #4257

+++ Tests/__integration_tests__/requestApproval.integration.spec.js
@@ -122,7 +122,7 @@ describe('支払依頼画面のインテグレーションテスト', () => {
 
       await page.click('#informationTab > table > tbody > tr:nth-child(1) > td.text-center.display-row-td > a')
 
-      await page.waitForTimeout(1500)
+      await page.waitForTimeout(5000)
 
       redirectUrl = await page.evaluate(() => {
         return document.querySelector('#form > div.grouped-button > a.button.is-link').getAttribute('href')
@@ -150,7 +150,7 @@ describe('支払依頼画面のインテグレーションテスト', () => {
 
       await page.click('#informationTab > table > tbody > tr:nth-child(1) > td.text-center.display-row-td > a')
 
-      await page.waitForTimeout(1500)
+      await page.waitForTimeout(5000)
 
       await page.click('#form > div.grouped-button > a.button.is-link')
 
@@ -215,7 +215,7 @@ describe('支払依頼画面のインテグレーションテスト', () => {
 
       await page.click('#informationTab > table > tbody > tr:nth-child(1) > td.text-center.display-row-td > a')
 
-      await page.waitForTimeout(1500)
+      await page.waitForTimeout(5000)
 
       await page.click('#form > div.grouped-button > a.button.is-link')
 
@@ -239,7 +239,7 @@ describe('支払依頼画面のインテグレーションテスト', () => {
 
       await page.click('#informationTab > table > tbody > tr:nth-child(1) > td.text-center.display-row-td > a')
 
-      await page.waitForTimeout(1500)
+      await page.waitForTimeout(5000)
 
       await page.click('#form > div.grouped-button > a.button.is-link')
 
@@ -306,7 +306,7 @@ describe('支払依頼画面のインテグレーションテスト', () => {
 
       await page.click('#informationTab > table > tbody > tr:nth-child(1) > td.text-center.display-row-td > a')
 
-      await page.waitForTimeout(1500)
+      await page.waitForTimeout(5000)
 
       await page.click('#form > div.grouped-button > a.button.is-link')
 
@@ -330,7 +330,7 @@ describe('支払依頼画面のインテグレーションテスト', () => {
 
       await page.click('#informationTab > table > tbody > tr:nth-child(1) > td.text-center.display-row-td > a')
 
-      await page.waitForTimeout(1500)
+      await page.waitForTimeout(5000)
 
       await page.click('#form > div.grouped-button > a.button.is-link')
 
@@ -352,9 +352,9 @@ describe('支払依頼画面のインテグレーションテスト', () => {
       await page.setCookie(acCookies[0])
       await page.goto(`https://localhost:3000${redirectUrl}`)
 
-      await page.click('#form > div.grouped-button > a:nth-child(1)')
+      await page.click('#form > div.field.is-grouped.is-grouped-centered > div > div > div > a.button.mr-6')
 
-      await page.waitForTimeout(1500)
+      await page.waitForTimeout(5000)
 
       const inboxUrl = 'inbox/' + redirectUrl.split('/')[2]
 
@@ -383,7 +383,6 @@ describe('支払依頼画面のインテグレーションテスト', () => {
       expect(res.text).toMatch(/合計 円/i)
       expect(res.text).toMatch(/メッセージ/i)
       expect(res.text).toMatch(/仕訳情報設定へ/i)
-      expect(res.text).toMatch(/保存/i)
       expect(res.text).toMatch(/確認/i)
     })
 
@@ -405,7 +404,6 @@ describe('支払依頼画面のインテグレーションテスト', () => {
       expect(res.text).toMatch(/合計 円/i)
       expect(res.text).toMatch(/メッセージ/i)
       expect(res.text).toMatch(/仕訳情報設定へ/i)
-      expect(res.text).toMatch(/保存/i)
       expect(res.text).toMatch(/確認/i)
     })
 
@@ -509,7 +507,7 @@ describe('支払依頼画面のインテグレーションテスト', () => {
       // 承認ルート検索
       await page.click('#btnSearchApproveRoute')
 
-      await page.waitForTimeout(2000)
+      await page.waitForTimeout(3000)
 
       await page.click('#displayFieldApproveRouteResultBody > tr > td.btnSelect > a')
 
@@ -517,11 +515,6 @@ describe('支払依頼画面のインテグレーションテスト', () => {
 
       // メッセージ入力
       await page.type('#inputMsg', 'インテグレーションテスト')
-
-      // 保存ボタン押下
-      await page.click('#form > div.grouped-button > button')
-
-      await page.waitForTimeout(7000)
 
       // 支払依頼画面にredirectする。
       expect(page.url()).toBe(`https://localhost:3000${redirectUrl}`)
@@ -534,7 +527,7 @@ describe('支払依頼画面のインテグレーションテスト', () => {
       // 依頼ボタン押下
       await page.click('#btn-approval')
 
-      await page.waitForTimeout(3000)
+      await page.waitForTimeout(10000)
 
       // 一覧画面にredirectする。
       expect(await page.url()).toBe('https://localhost:3000/inboxList/1')
@@ -577,7 +570,7 @@ describe('支払依頼画面のインテグレーションテスト', () => {
         }
       })
 
-      if (requestId.length !== 0) {
+      if (requestId && requestId.length !== 0) {
         await db.Approval.destroy({ where: { requestId: requestId.requestId } })
         await db.RequestApproval.destroy({ where: { contractId: contract.contractId } })
       }
@@ -613,7 +606,7 @@ describe('支払依頼画面のインテグレーションテスト', () => {
 
       await page.click('#informationTab > table > tbody > tr:nth-child(1) > td.text-center.display-row-td > a')
 
-      await page.waitForTimeout(1500)
+      await page.waitForTimeout(5000)
 
       await page.click('#form > div.grouped-button > a.button.is-link')
 
@@ -637,7 +630,7 @@ describe('支払依頼画面のインテグレーションテスト', () => {
 
       await page.click('#informationTab > table > tbody > tr:nth-child(1) > td.text-center.display-row-td > a')
 
-      await page.waitForTimeout(1500)
+      await page.waitForTimeout(5000)
 
       await page.click('#form > div.grouped-button > a.button.is-link')
 
@@ -702,7 +695,7 @@ describe('支払依頼画面のインテグレーションテスト', () => {
 
       await page.click('#informationTab > table > tbody > tr:nth-child(1) > td.text-center.display-row-td > a')
 
-      await page.waitForTimeout(1500)
+      await page.waitForTimeout(5000)
 
       await page.click('#form > div.grouped-button > a.button.is-link')
 
@@ -726,7 +719,7 @@ describe('支払依頼画面のインテグレーションテスト', () => {
 
       await page.click('#informationTab > table > tbody > tr:nth-child(1) > td.text-center.display-row-td > a')
 
-      await page.waitForTimeout(1500)
+      await page.waitForTimeout(5000)
 
       await page.click('#form > div.grouped-button > a.button.is-link')
